fix(permissions): break circular dependency between permissions and evaluator
Split permissions.ts into permissions.ts (constants only) and resolve.ts
(resolve/buildPermissions functions that depend on evaluator). Fixes
circular import: permissions → evaluator → deny-map → permissions.
