+++ packages/server/lib/authz/index.ts
@@ -1,7 +1,8 @@
 export { authzMiddleware } from './middleware.js';
 export { authorize } from './authorize.js';
 export { evaluator } from './evaluator.js';
-export { buildPermissions, permissions, resolve } from './permissions.js';
+export { permissions } from './permissions.js';
+export { buildPermissions, resolve } from './resolve.js';
 export { ROLE_DENY_MAP } from './deny-map.js';
 export { envScope, registerPermission } from './resolvers.js';
 export type { Permission, PermissionEvaluator } from './types.js';
